database.py

from sqlalchemy.orm import sessionmaker, declarative_base
import os
from dotenv import load_dotenv
from sqlalchemy import create_engine

load_dotenv()

DATABASE_URL = os.getenv('DATABASE_URL')


engine = create_engine(
    DATABASE_URL
)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)


# Dependency

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()



